mysite/requirements.txt
# Dev environment
pip
autopep8

# App
asgiref==3.7.2
Django==5.0
sqlparse==0.4.4
typing_extensions==4.9.0
pillow==10.4.0
face-recognition==1.3.0
numpy==1.26.3
opencv-python==4.9.0.80
djangorestframework
markdown
django-filter
django-cors-headers==4.3.1
scikit-learn==1.5.1
#deepface
djangorestframework-simplejwt